requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via typer
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
anyio==4.12.1
    # via httpx
attrs==25.4.0
    # via aiohttp
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.5
    # via requests
chess==1.11.2
    # via python-chess
click==8.3.1
    # via typer
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==4.6.1
    # via chess-classifier (pyproject.toml)
dill==0.4.0
    # via
    #   datasets
    #   multiprocess
duckdb==1.4.4
    # via chess-classifier (pyproject.toml)
filelock==3.25.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
fonttools==4.61.1
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.2.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via httpcore
hf-xet==1.3.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   datasets
    #   huggingface-hub
huggingface-hub==1.6.0
    # via datasets
hydra-core==1.3.2
    # via chess-classifier (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via chess-classifier (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.18
    # via datasets
networkx==3.6.1
    # via torch
numpy==2.4.2
    # via
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   xgboost
omegaconf==2.3.0
    # via hydra-core
packaging==26.0
    # via
    #   datasets
    #   huggingface-hub
    #   hydra-core
    #   matplotlib
pandas==3.0.1
    # via datasets
pillow==12.1.1
    # via matplotlib
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyarrow==23.0.1
    # via datasets
pygments==2.19.2
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-chess==1.999
    # via chess-classifier (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pyyaml==6.0.3
    # via
    #   datasets
    #   huggingface-hub
    #   omegaconf
requests==2.32.5
    # via datasets
rich==14.3.3
    # via typer
scikit-learn==1.8.0
    # via chess-classifier (pyproject.toml)
scipy==1.17.1
    # via
    #   scikit-learn
    #   xgboost
setuptools==82.0.0
    # via torch
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.10.0
    # via chess-classifier (pyproject.toml)
tqdm==4.67.3
    # via
    #   datasets
    #   huggingface-hub
typer==0.24.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.6.3
    # via requests
xgboost==3.2.0
    # via chess-classifier (pyproject.toml)
xxhash==3.6.0
    # via datasets
yarl==1.23.0
    # via aiohttp
